+++ src/firebolt/async_db/util.py
@@ -11,8 +11,6 @@ from firebolt.utils.exception import (
 )
 from firebolt.utils.urls import GATEWAY_HOST_BY_ACCOUNT_NAME
 
-ENGINE_STATUS_RUNNING_LIST = ["Running", "ENGINE_STATE_RUNNING"]
-
 
 async def _get_system_engine_url(
     auth: Auth,
