+++ src/firebolt/common/constants.py
@@ -1,3 +1,6 @@
 KEEPALIVE_FLAG: int = 1
 KEEPIDLE_RATE: int = 60  # seconds
 DEFAULT_TIMEOUT_SECONDS: int = 60
+
+# Running statuses in infromation schema
+ENGINE_STATUS_RUNNING_LIST = ["Running", "ENGINE_STATE_RUNNING"]
