+++ src/firebolt/db/connection.py
@@ -10,9 +10,12 @@ from httpx import Timeout
 from firebolt.client import DEFAULT_API_URL, Client, ClientV1, ClientV2
 from firebolt.client.auth import Auth
 from firebolt.common.base_connection import BaseConnection
-from firebolt.common.constants import DEFAULT_TIMEOUT_SECONDS
+from firebolt.common.constants import (
+    DEFAULT_TIMEOUT_SECONDS,
+    ENGINE_STATUS_RUNNING_LIST,
+)
 from firebolt.db.cursor import Cursor, CursorV1, CursorV2
-from firebolt.db.util import ENGINE_STATUS_RUNNING_LIST, _get_system_engine_url
+from firebolt.db.util import _get_system_engine_url
 from firebolt.utils.exception import (
     ConfigurationError,
     ConnectionClosedError,
